requirements.txt
Django
djangorestframework
markdown
django-filter
django-cors-headers
mysqlclient
djangorestframework-api-key
